Add openssh-client to v1.33
Since we still maintain this previous version, we have to support openssh (as in v2).
We do not want to tag v1.X as latest since there is a version2
Since the lastest tag is always the latest version pushed, we need to push back v2 after previous version release.